app/backend/appengine/requirements.txt
Flask==1.1.2
google_api_python_client==1.12.8

